+++ cli/cli/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - create default `.gitignore` file during `beam init` if none is present
+- `beam version` commands to manage CLI version install
 
 ### Fixed